requirements.txt
g2p-en ~= 2.1
inflect ~= 4.1
librosa ~= 0.7
matplotlib ~= 3.2
numba
numpy ~= 1.19
pypinyin~=0.39
pyworld ~= 0.2
PyYAML
scikit-learn
scipy ~= 1.5
soundfile~=0.10
tensorboard ~= 2.2
tgt ~= 1.4
torch ~= 1.7
tqdm
unidecode ~= 1.1
seaborn